requirements.txt
asgiref==3.2.7
certifi==2020.6.20
chardet==3.0.4
Django==3.0.6
djangorestframework==3.11.0
idna==2.10
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
unittest-xml-reporting==3.0.2
urllib3==1.25.9